+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The default value for `draw_bold_text_with_bright_colors` is now `false`
 - Mirror OSC query terminators instead of always using BEL
 - Increased Beam, Underline, and Hollow Block cursors' line widths
-
+- Added kinetic scrolling. New Setting: `Scrolling/kinetic_friction` (0 for no Kinetic scrolling)
 ### Fixed
 
 - Incorrect default config path in `--help` on Windows and macOS